+++ app/src/main/java/com/project/medibox/medication/resources/CreateReminderResource.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class CreateReminderResource(
     @SerializedName("createdDate")
     var createdDateString: String,
-    var pills: Int?,
+    var pills: Short?,
     @SerializedName("endDate")
     var endDateString: String?,
     var medicineId: Long,
